# Optimize for a fair comparison