The comparison was not fair, so I optimized the workload so that PostgreSQL and MongoDB are comparable:
- MongoDB does a batch insert, so I changed the row-by-row in PostgreSQL to insert by batch of 1000
- No indexes are created on MongoDB, but an expensive GIN index was created on PostgreSQL, not useful for those queries. I've set the same indexes that can serve the queries. On scalar values, expression index on a JSON path is better and covers range queries (GIN doesn't)
- PostgreSQL is lazy and delays some cleanup to later (vacuum) so it's not fair to compare on a very short run. I changed to run 100 times.

Here are some numbers run on Gitpod:
```
📊 Benchmark run 50/100
✅ Generated 10000 dummy records for benchmarking
📝 PostgreSQL Insert Time: 162.071234ms
📝 MongoDB Insert Time: 152.706244ms
🔍 PostgreSQL Read Time: 753.57µs
🔍 MongoDB Read Time: 653.18µs

2025/06/25 15:40:34 /workspace/Jsonb-Postgresql/internal/db/postgresql.go:121 SLOW SQL >= 200ms
[13584.321ms] [rows:732708]   
                UPDATE json_data  
                SET data = jsonb_set(data, '{address,city}', to_jsonb('BenchCity'::text), false)  
                WHERE (data->>'age')::int > 30
✏️ PostgreSQL Update Time: 13.584429055s
✏️ MongoDB Update Time: 4.375644751s
🗑️ PostgreSQL Delete Time: 1.26462ms
🗑️ MongoDB Delete Time: 621.82µs
```
MongoDB performs faster on insert, read, and delete operations, with a significant advantage in update tasks.