@faeiiz 2025-07-01
I was also looking into postgresql being slow coz of heavy index on all the data might actually make the read slow. 
I also thought of switching of the WAL ( write ahead logging ) in postgres to see if it actually make it a bit better. 

and Thank you. I didn't know this -> No indexes are created on MongoDB, but an expensive GIN index was created on PostgreSQL, not useful for those queries. I've set the same indexes that can serve the queries. On scalar values, expression index on a JSON path is better and covers range queries (GIN doesn't).